api/requirements.txt
fastapi==0.104.1
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6 
SQLAlchemy==2.0.23 